**Summary**

This PR provides various performance optimizations to the evaluation of the SNAP potential within the Kokkos package. One optimization is taking advantage of the symmetry inherent to the Wigner U-matrices (ulist, ulisttot) and derivatives thereof (dulist), asymptotically reducing memory overheads by half. Taking advantage of this symmetry reduces the number of atomic adds into ulisttot. In addition, for the GPU codepath only, the Cayley-Klein parameters and derivates thereof are now pre-computed in a way that better exploits parallelism than the current implementation, more than amortizing the overheads of additional memory IO.

These optimizations offer a performance improvement, or at least performance parity (for very small systems), for both the CPU and GPU.

Correctness was verified against all flavors of SNAP: single atom (W), multi-atom w/ and w/out chemsnap (InP and WBe, respectively), and quadratic (hacking Mo_Zuo potential data into Mo_Chen input file). Performance notes below. For robustness, I note that master is specifically commit `ac43f8f6857153b76cd3e65b388f637ca4ca6f19` at time of filing. Performance measurements are from a week previous, but there have been no substantive changes to the relevant routines in that time.

Performance improvements on V100-16GB + x86:

Input 				file | Timesteps/sec 				master | Timesteps/sec 				PR | Perf 				improvement | Comment
-- | -- | -- | -- | --
2J8 				benchmark | 275 | 337 | 1.23x | Benchmark 				problem size (2000 atoms)
2J14 				benchmark | 50.2 | 55.3 | 1.10x | ECP 				KPP problem size (2000 atoms)
in.snap.W.2940 | 1600 | 1600 | unchanged | /examples/snap/ single 				atom (128 atoms)
in.snap.WBe.PRB2019 | 1550 | 1560 | 1.01x | /examples/snap/ multi-atom, 				no chem (128 atoms)
in.snap.InP.JCPA2020 | 470 | 524 | 1.11x | /examples/snap/ multi-atom, 				yes chem (512 atoms)

Performance changes on 10 core Broadwell (used KOKKOS_ARCH=BDW appropriately)

Input 				file | Timesteps/sec 				master | Timesteps/sec 				PR | Perf 				change | Comment
-- | -- | -- | -- | --
2J8 				benchmark | 6.81 | 7.68 | 1.13x | Benchmark 				problem size (2000 atoms)
2J14 				benchmark | 0.759 | 0.786 | 1.04x | ECP 				KPP problem size (2000 atoms)
in.snap.W.2940 | 115 | 115 | Unchanged | /examples/snap/ single 				atom (128 atoms)
in.snap.WBe.PRB2019 | 112 | 116 | 1.04x | /examples/snap/ multi-atom, 				no chem (128 atoms)
in.snap.InP.JCPA2020 | 24.9 | 27.9 | 1.12x | /examples/snap/ multi-atom, 				yes chem (512 atoms)



**Related Issues**

This was appropriately reconciled with the changes in #2275 for HIP support.

**Author(s)**

Evan Weinberg (NVIDIA). eweinberg [at] nvidia.com preferred, evansweinberg [at] gmail.com for "long-lived".

**Licensing**

By submitting this pull request, I agree, that my contribution will be included in LAMMPS and redistributed under either the GNU General Public License version 2 (GPL v2) or the GNU Lesser General Public License version 2.1 (LGPL v2.1).

**Backward Compatibility**

No changes.

**Implementation Notes**

- As part of the GPU codepath, precomputes the Cayley-Klein parameters and derivatives thereof. This is baked into the neighbor table build as it removes the need for the `wj`, `rij`, data structures. This offers improved parallelism for this computation, which more than amortizes the memory loads/stores in `ComputeUi` and `ComputeFusedDeidrj` routines.
- (Effectively) halves the size of the `ulisttot` and `ylist` data structures. In the former case, this is taking advantage of the symmetry in the Wigner U matrices. In the latter case, it’s taking advantage of how the bispectrum elements in `ylist` are contracted with derivatives of the Wigner U matrices, which also obey symmetries, eliminating the need for all of the `ylist` elements when done correctly.
- CPU only: asymptotically halves the size of the `ulist` and `dulist` data structures by using a compressed (“cached”) data structure. This takes inspiration from the scratch memory data layout on the GPU, which needs an extra row of the even-ranked U matrices relative to the amount required using symmetry alone.
- Eliminates half of the `atomicAdd` calls in `ComputeUi` by the virtue of using the symmetrized data layout.
- Added `Kokkos::ViewAllocateWithoutInitializing` to allocations in `grow_rij`. This made a performance difference of note for the Cayley-Klein parameter structures, but was added elsewhere as a general diagnostic tool---unnecessarily zero-initialized data hides bugs, as I’ve learned the hard way...
- Added several comments throughout the code with the intent of documenting the modified workflows and explaining the symmetrized data layouts.

**Post Submission Checklist**

- [x] The feature or features in this pull request is complete
- [ ] Licensing information is complete
- [x] Corresponding author information is complete
- [ ] The source code follows the LAMMPS formatting guidelines
- [x] The feature has been verified to work with the conventional build system
- [x] The feature has been verified to work with the CMake based build system

**Further Information, Files, and Links**

None.