@akohlmey 2020-08-12
While you guys are working on that part of the code...
When compiling with the OpenMP backend, I get a whole bunch of rather annoying compiler warnings as shown below and repeated several times. Can this pragma be properly ifdef'd, so that only the compiler than is *supposed* to see and process it will get to see it?

```
/home/akohlmey/compile/lammps/src/KOKKOS/sna_kokkos_impl.h:489: warning: ignoring #pragma unroll  [-Wunknown-pragmas]
  489 |       #pragma unroll
      | 
/home/akohlmey/compile/lammps/src/KOKKOS/sna_kokkos_impl.h:496: warning: ignoring #pragma unroll  [-Wunknown-pragmas]
  496 |         #pragma unroll
      | 
/home/akohlmey/compile/lammps/src/KOKKOS/sna_kokkos_impl.h:654: warning: ignoring #pragma unroll  [-Wunknown-pragmas]
  654 |       #pragma unroll
      | 
/home/akohlmey/compile/lammps/src/KOKKOS/sna_kokkos_impl.h:661: warning: ignoring #pragma unroll  [-Wunknown-pragmas]
  661 |         #pragma unroll
      | 
```