@stanmoore1 2020-08-12
@akohlmey sure, should be easy to fix