@stanmoore1 2020-08-17
@weinbe2 This is not directly related to your changes in this PR, but in several places, we have a `vector_size` of 1 (i.e. see https://github.com/lammps/lammps/blob/master/src/KOKKOS/pair_snap_kokkos_impl.h#L234) but still have a `ThreadVectorRange` that is not doing anything, (i.e. https://github.com/lammps/lammps/blob/master/src/KOKKOS/sna_kokkos_impl.h#L291). Is there any reason to keep this in the code?