@weinbe2 2020-08-17
> @weinbe2 This is not directly related to your changes in this PR, but in several places, we have a `vector_size` of 1 (i.e. see https://github.com/lammps/lammps/blob/master/src/KOKKOS/pair_snap_kokkos_impl.h#L234) but still have a `ThreadVectorRange` that is not doing anything, (i.e. https://github.com/lammps/lammps/blob/master/src/KOKKOS/sna_kokkos_impl.h#L291). Is there any reason to keep this in the code?

I just gave it a quick skim, they're all the CPU routines, and I don't see any that are worth keeping---either it would take a non-trivial amount of work to make them vectorize (if we assumed `ThreadVectorRange` did do a good job of generating `SIMD` instructions as it were), or you'd probably just rip them out once you start experimenting with `Kokkos::SIMD` anyway.