@stanmoore1 2020-08-17
@weinbe2 On Summit I see about 20% improvement for `2J8` and 7% improvement for `2J14`.