@stanmoore1 2020-08-17
@weinbe2 FYI I also only see 306 timesteps/s for 2J8 on Summit, vs your 337 timesteps/s. Probably `Power9` vs `x86 `difference?