@weinbe2 2020-08-17
@stanmoore1 which version of CUDA are you using on Summit? My numbers were from CUDA 11.0, I'll admit I haven't looked at CUDA 10.1 numbers (which I think is still the Summit default?)

Given how much time SNAP spends in kernels I'd be surprised if `Power9` vs `x86` alone gave a ~10% perf difference