@stanmoore1 2020-08-17
@weinbe2 , using CUDA 11.0 actually shows a slight performance regression over 10.1, which is not good. I used the same node on Summit and averaged over 6 runs for both versions:

| CUDA version| Perf|
|---|---|
|10.1.243|311.3 timesteps/s|
|11.0.3|304.7 timesteps/s|

This result is quite reproducible. So now we have two performance mysteries (10% slowdown on x86 vs Summit, and CUDA versions) :(