@weinbe2 2020-08-17
Oi, okay. I can dig in tomorrow. I should be able to close the 10.1.243 vs 11.0.3 perf gap. x86 vs Power9 will be more interesting.