@weinbe2 2020-08-17
@stanmoore1 What are the timesteps/sec for 2J14 on Power9?