@stanmoore1 2020-08-17
Here are the 2J14 results. I'm averaging 6 concurrent runs, one on each GPU. 

| CUDA version| 2J14 Perf|
|---|---|
|10.1.243|56.1 timesteps/s|
|11.0.3|55.2 timesteps/s|