@stanmoore1 2020-08-17
@weinbe2 with CUDA 10.1.243, I see 1.25x and 1.09x speedup for 2J8 and 2J14 respectively, which matches your speedups.