@stanmoore1 2020-08-18
I spot checked the numerics and verified that this PR still gives the correct answer.