@akohlmey 2020-08-18
> @akohlmey I'm working on the `#pragma unroll` warning issue now, will submit a new PR.

you could add those changes to PR #2289