@stanmoore1 2020-08-18
OK will do.